fix(store/sqlite): move multi-entity tests out of store_test.go
store_test.go was accumulating cross-entity integration tests that
do not belong to its infrastructure-testing role. Move four tests:

  TestPatchDomainUserResource
  TestDelete_userGroupResource_success
  TestDelete_notFound
  TestPatch_emptyInvalid_notFound

into a new store_multi_entity_test.go, keeping store_test.go focused
on store-level helpers (wrapConstraintError, closedDB, list helpers).
